Adds get all destinations endpoint(https://trello.com/c/O7uhdVJ2)

+++ migrations/2023-05-21-155745_add_destinations/up.sql
@@ -5,8 +5,8 @@ CREATE TABLE destinations (
   visibility VARCHAR(255) NOT NULL,
   isReviewed BOOLEAN NOT NULL,
   name VARCHAR(255) NOT NULL,
-  latitude FLOAT NOT NULL,
-  longitude FLOAT NOT NULL
+  latitude DOUBLE NOT NULL,
+  longitude DOUBLE NOT NULL
 );
 
 ALTER TABLE destinations ADD FOREIGN KEY (ownerId) REFERENCES users(id);
