+++ src/db/schema.rs
@@ -7,8 +7,8 @@ diesel::table! {
         visibility -> Varchar,
         isReviewed -> Bool,
         name -> Varchar,
-        latitude -> Float,
-        longitude -> Float,
+        latitude -> Double,
+        longitude -> Double,
     }
 }
 
@@ -24,4 +24,7 @@ diesel::table! {
 
 diesel::joinable!(destinations -> users (ownerId));
 
-diesel::allow_tables_to_appear_in_same_query!(destinations, users,);
+diesel::allow_tables_to_appear_in_same_query!(
+    destinations,
+    users,
+);
